Enabled build time rendering of article pages
This was a bit crap. To enable this I had to make sure that every file I touched in the next.config.js file was commonJS, because these aren't transpiled by NextJs. This also meant that I needed to duplicate some of the code (changing it to CommonJs) from blogApi as this was written in es6 and typescript. I also needed to make sure that export path map didn't export any of the default pages that needed a slug to function. It's a bit of duplicate work but I think it's a one time cost.