Fixed initial width issue on mobile and removed font-awesome
I forgot a meta tag tells the browser to match the browser window width to the width of the screen. For whatever reason the default is to make the browser window much bigger, make everything look zoomed out. Also didn't need font-awesome and it wasn't worth the hassle to leave it in due to the npmrc injected environment variable